+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# Version 0.2.0
+
+- Alembic Migrations
+- Support unknown arguments in management commands
+- `py.test --args='xs'` -> `py.test -xs`
+- Optional SSL support for SMTP
+- Additional documentation
+- Tightening up manage.py import/structure
+
 # Version 0.1.2
 
 - Async SMTP Support
+
+# Version 0.1.1
+
+- Admin API: put -> patch
+- Modern string formatting
+- Added sort, limit, offset to admin search
+
+# Version 0.1.0
+
+- Initial release
+- Basic structure and core features of server
+- Management commands
+- New project/app scaffolding
+- Basic admin API
+- User Auth
+- CSRF Protection
+- Async Postgresql support
+- SQLAlchemy Core support
+- Unit tests via py.test